Make pterm's default window title more sensible.
In the rework last month, term_setup_window_titles made a centralised
decision about whether to make the default window title just the
application name, or the hostname and application name separated by a
dash. Unfortunately it based that decision on whether the hostname was
NULL, whereas I'd meant to include a _empty_ hostname as not worth
printing.

As a result, the default window title for pterm became " - pterm", not
just "pterm". Easily fixed.